#1034 ignore case when searching http headers by name
