+++ src/test/java/com/codeborne/selenide/impl/DownloadFileWithProxyServerTest.java
@@ -3,8 +3,8 @@ package com.codeborne.selenide.impl;
 import com.codeborne.selenide.Browser;
 import com.codeborne.selenide.DriverStub;
 import com.codeborne.selenide.SelenideConfig;
+import com.codeborne.selenide.proxy.DownloadedFile;
 import com.codeborne.selenide.proxy.FileDownloadFilter;
-import com.codeborne.selenide.proxy.FileDownloadFilter.DownloadedFile;
 import com.codeborne.selenide.proxy.SelenideProxyServer;
 import com.google.common.collect.ImmutableSet;
 import org.assertj.core.api.WithAssertions;
